UI interactions work
The keys are mapped to the correct inputs and we should be able to update each agents data with the menus. The script for tasks is started but the current implementation only moves it a little when you hit C on movenow. I need to expand the tasks more to make the movement work